requirements.txt
# Web框架
flask>=2.0.0
werkzeug>=2.0.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.0

# 数据库
mysql-connector-python>=8.0.0
sqlite3

# 计算机视觉和图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0
facenet-pytorch>=2.5.0

# TensorFlow相关（备用支持）
tensorflow>=2.6.0
tensorflow-hub>=0.12.0
tensorflow-datasets>=4.4.0

# 传统人脸识别库（备用）
face-recognition>=1.3.0
cmake>=3.18.0
dlib>=19.22.0

# 科学计算
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0

# 进度条和工具
tqdm>=4.62.0
requests>=2.26.0

# 数据库连接池（可选）
sqlalchemy>=1.4.0
pymysql>=1.0.0